put everything in the result block
starting on day8 with this, it's mostly just stylistic,
so I'm not super worried about going back and fixing everything.

At some point, might be worth doing the dataloading automatically
so that #parse! becomes a call inside Result that hands off to TaskData
and then exposes the data inside the block, save a little boilerplate.